+++ src/main/java/io/oken1/modules/mug/service/ContentService.java
@@ -13,6 +13,6 @@ public interface ContentService extends IService<ContentEntity> {
 
     Object insertGameContent(String startDate, String endDate);
 
-    Object updateContent(int aid, String oldContent, String newType, String newContent);
+    Object updateContent(Long aid, String oldContent, String newType, String newContent);
 }
